Put all exception classes under a directory Put each connection string exception in an individual file
