+++ src/EventStore.Client/Exceptions/AccessDeniedException.cs
