+++ src/EventStore.Client/Exceptions/ConnectionString/ConnectionStringParseException.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace EventStore.Client {
+	public class ConnectionStringParseException : Exception {
+		public ConnectionStringParseException(string message) : base(message) { }
+	}
+}
