+++ src/EventStore.Client/Exceptions/ConnectionString/InvalidHostException.cs
@@ -1,0 +1,6 @@
+namespace EventStore.Client {
+	public class InvalidHostException : ConnectionStringParseException {
+		public InvalidHostException(string host)
+			: base($"Invalid host: '{host}'") { }
+	}
+}
