+++ src/EventStore.Client/Exceptions/ConnectionString/InvalidKeyValuePairException.cs
@@ -1,0 +1,6 @@
+namespace EventStore.Client {
+	public class InvalidKeyValuePairException : ConnectionStringParseException {
+		public InvalidKeyValuePairException(string keyValuePair)
+			: base($"Invalid key/value pair: '{keyValuePair}'") { }
+	}
+}
