+++ src/EventStore.Client/Exceptions/ConnectionString/InvalidSchemeException.cs
@@ -1,0 +1,6 @@
+namespace EventStore.Client {
+	public class InvalidSchemeException : ConnectionStringParseException {
+		public InvalidSchemeException(string scheme, string[] supportedSchemes)
+			: base($"Invalid scheme: '{scheme}'. Supported values are: {string.Join(",", supportedSchemes)}") { }
+	}
+}
