+++ src/EventStore.Client/Exceptions/ConnectionString/InvalidSettingException.cs
@@ -1,0 +1,5 @@
+namespace EventStore.Client {
+	public class InvalidSettingException : ConnectionStringParseException {
+		public InvalidSettingException(string message) : base(message) { }
+	}
+}
