+++ src/EventStore.Client/Exceptions/ConnectionString/NoSchemeException.cs
@@ -1,0 +1,6 @@
+namespace EventStore.Client {
+	public class NoSchemeException : ConnectionStringParseException {
+		public NoSchemeException()
+			: base("Could not parse scheme from connection string") { }
+	}
+}
