+++ src/EventStore.Client/Exceptions/DiscoveryException.cs
