+++ src/EventStore.Client/Exceptions/NotAuthenticatedException.cs
