+++ src/EventStore.Client/Exceptions/NotLeaderException.cs
