+++ src/EventStore.Client/Exceptions/RequiredMetadataPropertyMissingException.cs
