+++ src/EventStore.Client/Exceptions/ScavengeNotFoundException.cs
