+++ src/EventStore.Client/Exceptions/StreamDeletedException.cs
