+++ src/EventStore.Client/Exceptions/StreamNotFoundException.cs
