+++ src/EventStore.Client/Exceptions/UserNotFoundException.cs
