+++ src/EventStore.Client/Exceptions/WrongExpectedVersionException.cs
